@@ -11,12 +11,12 @@
         }
     ],
     "require": {
-        "php": "^7.3|^8.0",
-        "symfony/framework-bundle":  "^4.0|^5.2",
-        "symfony/yaml": "^4.0|^5.2"
+        "php": "^7.4|^8.0",
+        "symfony/framework-bundle":  "^4.4|^5.3",
+        "symfony/yaml": "^4.4|^5.3"
     },
     "require-dev": {
-        "friendsofphp/php-cs-fixer": "^1.1|^2.8",
+        "friendsofphp/php-cs-fixer": "^3.0",
         "symfony/test-pack": "*",
         "symfony/debug-pack": "*"
     },
