@@ -4,7 +4,7 @@ namespace Raneomik\WatchdogBundle\Watchdog\Model;
 
 interface WatchdogUnitInterface
 {
-    const UNIT_KEY_MAP = [
+    public const UNIT_KEY_MAP = [
       'compound',
       'date_time',
       'date',
